requirements_dev.txt
pip
bump2version
wheel
watchdog
flake8
tox
coverage
Sphinx
twine
Click
pyyaml
pytest
